pyproject.toml
[tool.poetry]
name = "mylife-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
mem0ai = "^0.1.2"
ollama = "^0.3.1"
groq = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
